refactor: move utils sources in their own directory
