+++ userspace/sysdig/utils/supported_events.cpp
